Introduce feature to create pull requests
Since master branch is protected we need to introduce more "cautious" approach to update swagger specs.
There is a command line tool called hub but it is not available for grabs for linux other than Fedora.
We grab the binaries and install it. Then use in slightly modified swagger document publisher