Add support for OpenAPI 3.1.0
Updated swagger-ui-dist to a version that supports OpenAPI 3.1.0.  Under the current version an error containing "The provided definition does not specify a valid version field" is displayed when attempting to view an OpenAPI 3.1.0 specification.

Also added configuration to opt out of anonymised installation stats gathering (Scarf) that was added in swagger-ui-dist 5.18.0.